fix[blog.views]: Fixed TemplateDoesNotExist error
Fixed the 'TemplateDoesNotExist' error showing up when opening the home page. The problem was that our Django app couldn't find the 'home.html' template. I made sure that 'index.html' is there and correctly named in the '[views.py](http://views.py/)' file. Now, the home page works fine.

On branch main
Your branch is up to date with 'origin/main'.
Changes to be committed:
	modified:   nederlearn/settings.py
Changes not staged for commit:
	modified:   blog/views.py